refactor(task-detail): extract domain hooks and presentational components
Decompose TaskDetail into useDiffState, useTerminalCache, useTaskViewMode,
and useTaskDetailComments hooks plus TaskDetailAgentView/DiffView components.
Preserves perTaskUiState and terminal LRU cache semantics byte-for-byte.
